moved everything into get-involved.tsx
