+++ src/styles/get-involved/number-scroll.module.scss
@@ -6,8 +6,7 @@
 
   padding: 50px;
   border-radius: 20px;
-  margin-top: 50px;
-  margin-bottom: 200px;
+  margin: 50px auto 200px;
   justify-self: center;
   max-width: 1200px;
 }
